pyproject.toml
[project]
name = "x-post"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "psutil>=7.1.3",
    "pygobject",
    "python-dotenv>=1.2.1",
    "tweepy>=4.16.0",
]
